ptr_GC requires function "take" to take ownership of an object Added music:wait command Changed time_t from double to float to remove warnings and decrease redundancy Contemplating over the editor... Testing out some audio features Added custom background color Added construct_list to the texture manager - comiles a list of anmes of textures Minor tweaks Issue: Need to rethink is_key_pressed and is_mouse_pressed   - They only return true once in one frame

+++ time.hpp
@@ -5,7 +5,7 @@
 
 namespace engine
 {
-typedef double time_t;
+typedef float time_t;
 
 // Universal time converter for clock
 class utime
